MIDI: Improve event timing accuracy
The stream generator now produces timestamps in seconds, double precison
floating point, and the player now operates with these timestamps. All
existing player modules have been adapted to the changes.

+++ Plugins/MIDI/MIDI/MIDIDecoder.h
@@ -36,9 +36,9 @@ class TSFPlayer;
 
 	double sampleRate;
 
-	long totalFrames;
-	long framesLength;
-	long framesFade;
+	double totalFrames;
+	double framesLength;
+	double framesFade;
 	long framesRead;
 
 	float outputBuffer[1024 * 2];
